Stub restrictions from helper
- DRYs up code that was used in both the integration and view tests.
