+++ test/integration/coronavirus_local_restrictions_test.rb
@@ -1,9 +1,11 @@
 require "integration_test_helper"
 require "gds_api/test_helpers/mapit"
+require_relative "../support/coronavirus_local_restrictions_helpers"
 
 class CoronavirusLocalRestrictionsTest < ActionDispatch::IntegrationTest
   include GdsApi::TestHelpers::Mapit
   include GdsApi::TestHelpers::ContentStore
+  include CoronavirusLocalRestrictionsHelpers
 
   describe "current restrictions" do
     it "displays the tier one restrictions" do
@@ -222,59 +224,5 @@ class CoronavirusLocalRestrictionsTest < ActionDispatch::IntegrationTest
     assert page.has_css?("meta[property='og:title'][content='#{I18n.t('coronavirus_local_restrictions.lookup.meta_title')}']", visible: false)
     assert page.has_css?("meta[name='description'][content='#{I18n.t('coronavirus_local_restrictions.lookup.meta_description')}']", visible: false)
     assert page.has_css?("link[rel='canonical'][href='http://www.example.com/find-coronavirus-local-restrictions']", visible: false)
-  end
-
-  def stub_local_restriction(postcode:,
-                             name: "Tatooine",
-                             gss: SecureRandom.alphanumeric(10),
-                             country_name: "England",
-                             current_alert_level: nil,
-                             future_alert_level: nil)
-    areas = [
-      {
-        "gss" => gss,
-        "name" => name,
-        "type" => "LBO",
-        "country_name" => country_name,
-      },
-    ]
-    stub_mapit_has_a_postcode_and_areas(postcode, [], areas)
-
-    current_restriction = if current_alert_level
-                            { "alert_level" => current_alert_level,
-                              "start_date" => 1.week.ago.to_date,
-                              "start_time" => "10:00" }
-                          end
-
-    future_restriction = if future_alert_level
-                           { "alert_level" => future_alert_level,
-                             "start_date" => 1.week.from_now.to_date,
-                             "start_time" => "10:00" }
-                         end
-
-    local_restriction = LocalRestriction.new(gss, {
-      "name" => name,
-      "restrictions" => [current_restriction, future_restriction].compact,
-    })
-    LocalRestriction.stubs(:find).with(gss).returns(local_restriction)
-  end
-
-  def stub_no_local_restriction(
-    postcode:,
-    name: "Tatooine",
-    gss: SecureRandom.alphanumeric(10),
-    country_name: "England"
-  )
-    areas = [
-      {
-        "gss" => gss,
-        "name" => name,
-        "type" => "LBO",
-        "country_name" => country_name,
-      },
-    ]
-    stub_mapit_has_a_postcode_and_areas(postcode, [], areas)
-
-    LocalRestriction.stubs(:find).with(gss).returns(nil)
   end
 end
